extract static PebbleUtils from WebTemlateToadlet for easier reuse
